% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lloomr-package.R
\docType{package}
\name{lloomr-package}
\alias{lloomr}
\alias{lloomr-package}
\title{lloomr: Concept Induction from Text with Large Language Models}
\description{
An R implementation of the LLooM concept induction algorithm (Lam et al., 2024, \doi{10.1145/3613904.3642830}) for extracting and applying interpretable, high-level concepts from text data using large language models. The pipeline distills documents into salient excerpts and bullet-point summaries, clusters them via text embeddings, asks an LLM to synthesize named concepts with inclusion criteria, reviews and deduplicates the concept set, and then scores every document against every concept. Translated from the Python package 'text_lloom'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zilinskyjan/lloomr}
  \item Report bugs at \url{https://github.com/zilinskyjan/lloomr/issues}
}

}
\author{
\strong{Maintainer}: Jan Zilinsky \email{zilinsky09@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Michelle Lam (Author of the original Python implementation (text_lloom)) [copyright holder]
}

}
\keyword{internal}
