% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{lloom_results}
\alias{lloom_results}
\title{Get the score results from a session}
\usage{
lloom_results(sess)
}
\arguments{
\item{sess}{A \code{\link[=lloom_session]{lloom_session()}} after \code{\link[=lloom_score]{lloom_score()}}.}
}
\value{
The long score tibble (see \code{\link[=score_concepts]{score_concepts()}}).
}
\description{
Get the score results from a session
}
\examples{
\dontrun{
score_df <- lloom_results(sess)
}
}
